move behavior from core to extensions including test class added sample with notification panel animation
