+++ bootstrap-core/pom.xml
@@ -42,12 +42,6 @@
         </dependency>
         
         <dependency>
-	    <groupId>org.webjars</groupId>
-	    <artifactId>animate.css</artifactId>
-	    <version>${animate.css.version}</version>
-	</dependency>
-
-        <dependency>
             <groupId>joda-time</groupId>
             <artifactId>joda-time</artifactId>
         </dependency>
