+++ bootstrap-samples/src/main/java/de/agilecoders/wicket/samples/pages/ExtensionsPage.html
@@ -261,6 +261,22 @@
                             </form>
                             <div wicket:id="spinnerFeedback"></div>
                         </section>
+                        <section id="animation">
+							<div class="page-header">
+                                <h1>Animation</h1>
+                            </div>
+
+                            <p>Just-add-water CSS animations with <a target="_blank" href="http://daneden.github.io/animate.css/">Animate CSS</a>.</p>
+
+                            <div class="bs-docs-example">
+                                <form role="form">
+                                    <div class="form-group">
+										<a class="btn btn-default" wicket:id="startAnimation"></a>
+                               		</div>
+                                </form>
+                            	<div wicket:id="animationFeedback"></div>
+                            </div>
+                        </section>
                     </div>
                 </div>
             </div>
